DataConsumer/requirements.txt
cassandra-driver==3.25.0
uuid==1.30
confluent-kafka==2.3.0
mkl-service==2.4.0
